Make new and save/save_file files for remote package
