+++ remote/remote.go
@@ -1,23 +1,18 @@
 package remote
 
 import (
-	"crypto/tls"
 	"fmt"
 	"io"
 	"net/http"
 	"strings"
 	"time"
 
-	"github.com/buildpacks/imgutil/layer"
-
 	"github.com/google/go-containerregistry/pkg/authn"
 	"github.com/google/go-containerregistry/pkg/name"
 	v1 "github.com/google/go-containerregistry/pkg/v1"
-	"github.com/google/go-containerregistry/pkg/v1/empty"
 	"github.com/google/go-containerregistry/pkg/v1/mutate"
 	"github.com/google/go-containerregistry/pkg/v1/remote"
 	"github.com/google/go-containerregistry/pkg/v1/remote/transport"
-	"github.com/google/go-containerregistry/pkg/v1/static"
 	"github.com/google/go-containerregistry/pkg/v1/tarball"
 	"github.com/google/go-containerregistry/pkg/v1/types"
 	"github.com/pkg/errors"
@@ -40,275 +35,6 @@ type Image struct {
 type registrySetting struct {
 	insecure           bool
 	insecureSkipVerify bool
-}
-
-// NewImage returns a new Image that can be modified and saved to a Docker daemon.
-func NewImage(repoName string, keychain authn.Keychain, ops ...ImageOption) (*Image, error) {
-	imageOpts := &options{}
-	for _, op := range ops {
-		if err := op(imageOpts); err != nil {
-			return nil, err
-		}
-	}
-
-	platform := defaultPlatform()
-	if (imageOpts.platform != imgutil.Platform{}) {
-		platform = imageOpts.platform
-	}
-
-	image, err := emptyImage(platform)
-	if err != nil {
-		return nil, err
-	}
-
-	ri := &Image{
-		keychain:            keychain,
-		repoName:            repoName,
-		image:               image,
-		addEmptyLayerOnSave: imageOpts.addEmptyLayerOnSave,
-		registrySettings:    imageOpts.registrySettings,
-	}
-
-	if imageOpts.prevImageRepoName != "" {
-		if err := processPreviousImageOption(ri, imageOpts.prevImageRepoName, platform); err != nil {
-			return nil, err
-		}
-	}
-
-	if imageOpts.baseImageRepoName != "" {
-		if err := processBaseImageOption(ri, imageOpts.baseImageRepoName, platform); err != nil {
-			return nil, err
-		}
-	}
-
-	imgOS, err := ri.OS()
-	if err != nil {
-		return nil, err
-	}
-	if imgOS == "windows" {
-		if err := prepareNewWindowsImage(ri); err != nil {
-			return nil, err
-		}
-	}
-
-	if imageOpts.createdAt.IsZero() {
-		ri.createdAt = imgutil.NormalizedDateTime
-	} else {
-		ri.createdAt = imageOpts.createdAt
-	}
-
-	return ri, nil
-}
-
-func getRegistry(repoName string, registrySettings map[string]registrySetting) registrySetting {
-	for prefix, r := range registrySettings {
-		if strings.HasPrefix(repoName, prefix) {
-			return r
-		}
-	}
-	return registrySetting{}
-}
-
-func processPreviousImageOption(ri *Image, prevImageRepoName string, platform imgutil.Platform) error {
-	reg := getRegistry(prevImageRepoName, ri.registrySettings)
-
-	prevImage, err := NewV1Image(prevImageRepoName, ri.keychain, WithV1DefaultPlatform(platform), WithV1RegistrySetting(reg.insecure, reg.insecureSkipVerify))
-	if err != nil {
-		return err
-	}
-
-	prevLayers, err := prevImage.Layers()
-	if err != nil {
-		return errors.Wrapf(err, "getting layers for previous image with repo name %q", prevImageRepoName)
-	}
-
-	ri.prevLayers = prevLayers
-
-	return nil
-}
-
-func processBaseImageOption(ri *Image, baseImageRepoName string, platform imgutil.Platform) error {
-	reg := getRegistry(baseImageRepoName, ri.registrySettings)
-
-	baseImage, err := NewV1Image(baseImageRepoName, ri.keychain, WithV1DefaultPlatform(platform), WithV1RegistrySetting(reg.insecure, reg.insecureSkipVerify))
-	if err != nil {
-		return err
-	}
-
-	ri.image = baseImage
-
-	return nil
-}
-
-func prepareNewWindowsImage(ri *Image) error {
-	// only append base layer to empty image
-	cfgFile, err := ri.image.ConfigFile()
-	if err != nil {
-		return err
-	}
-	if len(cfgFile.RootFS.DiffIDs) > 0 {
-		return nil
-	}
-
-	layerBytes, err := layer.WindowsBaseLayer()
-	if err != nil {
-		return err
-	}
-
-	windowsBaseLayer, err := tarball.LayerFromReader(layerBytes) // TODO: LayerFromReader is deprecated; LayerFromOpener or stream.NewLayer are suggested alternatives however the tests do not pass when they are used
-	if err != nil {
-		return err
-	}
-
-	image, err := mutate.AppendLayers(ri.image, windowsBaseLayer)
-	if err != nil {
-		return err
-	}
-
-	ri.image = image
-
-	return nil
-}
-
-// v1Options is used to configure the behavior when a v1.Image is created
-type v1Options struct {
-	platform        imgutil.Platform
-	registrySetting registrySetting
-}
-
-type V1ImageOption func(*v1Options) error
-
-// WithV1DefaultPlatform provides Architecture/OS/OSVersion defaults for the new v1.Image.
-func WithV1DefaultPlatform(platform imgutil.Platform) V1ImageOption {
-	return func(opts *v1Options) error {
-		opts.platform = platform
-		return nil
-	}
-}
-
-// WithV1RegistrySetting registers options to use when accessing images in a registry in order to construct a v1.Image.
-func WithV1RegistrySetting(insecure, insecureSkipVerify bool) V1ImageOption {
-	return func(opts *v1Options) error {
-		opts.registrySetting = registrySetting{
-			insecure:           insecure,
-			insecureSkipVerify: insecureSkipVerify,
-		}
-		return nil
-	}
-}
-
-// NewV1Image returns a new v1.Image
-func NewV1Image(baseImageRepoName string, keychain authn.Keychain, ops ...V1ImageOption) (v1.Image, error) {
-	imageOpts := &v1Options{}
-	for _, op := range ops {
-		if err := op(imageOpts); err != nil {
-			return nil, err
-		}
-	}
-
-	platform := defaultPlatform()
-	if (imageOpts.platform != imgutil.Platform{}) {
-		platform = imageOpts.platform
-	}
-
-	reg := registrySetting{}
-	if (imageOpts.registrySetting != registrySetting{}) {
-		reg = imageOpts.registrySetting
-	}
-
-	baseImage, err := newV1Image(keychain, baseImageRepoName, platform, reg)
-	if err != nil {
-		return nil, err
-	}
-	return baseImage, nil
-}
-
-func newV1Image(keychain authn.Keychain, repoName string, platform imgutil.Platform, reg registrySetting) (v1.Image, error) {
-	ref, auth, err := referenceForRepoName(keychain, repoName, reg.insecure)
-	if err != nil {
-		return nil, err
-	}
-
-	v1Platform := v1.Platform{
-		Architecture: platform.Architecture,
-		OS:           platform.OS,
-		OSVersion:    platform.OSVersion,
-	}
-
-	opts := []remote.Option{remote.WithAuth(auth), remote.WithPlatform(v1Platform)}
-	// #nosec G402
-	if reg.insecureSkipVerify {
-		opts = append(opts, remote.WithTransport(&http.Transport{
-			TLSClientConfig: &tls.Config{
-				InsecureSkipVerify: true,
-			},
-		}))
-	} else {
-		opts = append(opts, remote.WithTransport(http.DefaultTransport))
-	}
-
-	var image v1.Image
-	for i := 0; i <= maxRetries; i++ {
-		time.Sleep(100 * time.Duration(i) * time.Millisecond) // wait if retrying
-		image, err = remote.Image(ref, opts...)
-		if err != nil {
-			if err == io.EOF && i != maxRetries {
-				continue // retry if EOF
-			}
-			if transportErr, ok := err.(*transport.Error); ok && len(transportErr.Errors) > 0 {
-				switch transportErr.StatusCode {
-				case http.StatusNotFound, http.StatusUnauthorized:
-					return emptyImage(platform)
-				}
-			}
-			if strings.Contains(err.Error(), "no child with platform") {
-				return emptyImage(platform)
-			}
-			return nil, errors.Wrapf(err, "connect to repo store %q", repoName)
-		}
-		break
-	}
-
-	return image, nil
-}
-
-func emptyImage(platform imgutil.Platform) (v1.Image, error) {
-	cfg := &v1.ConfigFile{
-		Architecture: platform.Architecture,
-		OS:           platform.OS,
-		OSVersion:    platform.OSVersion,
-		RootFS: v1.RootFS{
-			Type:    "layers",
-			DiffIDs: []v1.Hash{},
-		},
-	}
-
-	return mutate.ConfigFile(empty.Image, cfg)
-}
-
-func defaultPlatform() imgutil.Platform {
-	return imgutil.Platform{
-		OS:           "linux",
-		Architecture: "amd64",
-	}
-}
-
-func referenceForRepoName(keychain authn.Keychain, ref string, insecure bool) (name.Reference, authn.Authenticator, error) {
-	var auth authn.Authenticator
-	opts := []name.Option{name.WeakValidation}
-	if insecure {
-		opts = append(opts, name.Insecure)
-	}
-	r, err := name.ParseReference(ref, opts...)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	auth, err = keychain.Resolve(r.Context().Registry)
-	if err != nil {
-		return nil, nil, err
-	}
-	return r, auth, nil
 }
 
 func (i *Image) Label(key string) (string, error) {
@@ -724,78 +450,6 @@ func findLayerWithSha(layers []v1.Layer, diffID string) (v1.Layer, error) {
 		}
 	}
 	return nil, fmt.Errorf("previous image did not have layer with diff id %q", diffID)
-}
-
-func (i *Image) Save(additionalNames ...string) error {
-	return i.SaveAs(i.Name(), additionalNames...)
-}
-
-func (i *Image) SaveAs(name string, additionalNames ...string) error {
-	var err error
-
-	allNames := append([]string{name}, additionalNames...)
-
-	i.image, err = mutate.CreatedAt(i.image, v1.Time{Time: i.createdAt})
-	if err != nil {
-		return errors.Wrap(err, "set creation time")
-	}
-
-	cfg, err := i.image.ConfigFile()
-	if err != nil {
-		return errors.Wrap(err, "get image config")
-	}
-	cfg = cfg.DeepCopy()
-
-	layers, err := i.image.Layers()
-	if err != nil {
-		return errors.Wrap(err, "get image layers")
-	}
-	cfg.History = make([]v1.History, len(layers))
-	for j := range cfg.History {
-		cfg.History[j] = v1.History{
-			Created: v1.Time{Time: i.createdAt},
-		}
-	}
-
-	cfg.DockerVersion = ""
-	cfg.Container = ""
-	i.image, err = mutate.ConfigFile(i.image, cfg)
-	if err != nil {
-		return errors.Wrap(err, "zeroing history")
-	}
-
-	if len(layers) == 0 && i.addEmptyLayerOnSave {
-		empty := static.NewLayer([]byte{}, types.OCILayer)
-		i.image, err = mutate.AppendLayers(i.image, empty)
-		if err != nil {
-			return errors.Wrap(err, "empty layer could not be added")
-		}
-	}
-
-	var diagnostics []imgutil.SaveDiagnostic
-	for _, n := range allNames {
-		if err := i.doSave(n); err != nil {
-			diagnostics = append(diagnostics, imgutil.SaveDiagnostic{ImageName: n, Cause: err})
-		}
-	}
-	if len(diagnostics) > 0 {
-		return imgutil.SaveError{Errors: diagnostics}
-	}
-
-	return nil
-}
-
-func (i *Image) SaveFile() (string, error) {
-	return "", errors.New("not yet implemented")
-}
-
-func (i *Image) doSave(imageName string) error {
-	reg := getRegistry(i.repoName, i.registrySettings)
-	ref, auth, err := referenceForRepoName(i.keychain, imageName, reg.insecure)
-	if err != nil {
-		return err
-	}
-	return remote.Write(ref, i.image, remote.WithAuth(auth))
 }
 
 func (i *Image) Delete() error {
