+++ remote/save_file.go
@@ -1,0 +1,9 @@
+package remote
+
+import (
+	"errors"
+)
+
+func (i *Image) SaveFile() (string, error) {
+	return "", errors.New("not yet implemented")
+}
